third_party/blink/web_tests/platform/generic/fast/forms/fieldset/legend-form-expected.txt

This test checks the form attribute of the legend element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



- Ensures that the form attribute of legend element depends on whether its parent is a fieldset element or not.
PASS legendElement1.form is owner
PASS legendElement2.form is null
PASS legendElement3.form is null


- Ensures that the legend's form attribute points its parent fieldset's form owner even if the element is within another form element.
PASS fieldsetElement.form is owner
PASS legendElement.form is owner


- Ensures whether the form owner is set correctly for the legend when the value of form attribute of its parent fieldset changed.
PASS fieldsetElement.form is form1
PASS legendElement.form is form1
PASS fieldsetElement.form is form2
PASS legendElement.form is form2
PASS successfullyParsed is true

TEST COMPLETE
